requirements.txt
orjson~=3.10.7
fastapi~=0.115.6
sentence-transformers~=3.3.1
python-dotenv~=0.21.0
psutil~=5.9.0
httpx~=0.27.0
motor~=3.6.0
pydantic~=2.10.3